+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [3.0.0]
+### Added
+- Log more information about errors.
+- Rate limiting
+- Retry logic for Anthropic
+
+### Fixed
+- Many of the interrogees were broken.
+
 ## [2.0.0]
 ### Added
 - record metrics to postgres for analysis